refactor: クラスベースからfunctionベースのサービス層に変更
- UserServiceクラスをexported functionsに変更
- executeQueryとfetchQueryを共通関数として/src/database/query.tsに分離
- DatabaseErrorを共通クエリモジュールで定義
- CLAUDE.mdにfunction-based servicesのガイドラインを追加
